+++ projects/_03_independent_cqc/unittest_data/ind_cqc_test_file_schemas.py
@@ -841,7 +841,7 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsSchemas:
         [
             *calculate_difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles_schema,
             StructField(
-                IndCQC.difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles,
+                IndCQC.difference_estimate_filled_posts_and_from_all_job_roles,
                 FloatType(),
                 False,
             ),