+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_data.py
@@ -2276,13 +2276,11 @@ class EstimateFilledPostsByJobRoleEstimateUtilsData:
         ),
     ]
 
-    expected_calculate_difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles_rows = [
-        (0, 10.0, MainJobRoleLabels.care_worker, 5.0, 10.0, 0.0),
+    expected_calc_diff_estimate_filled_posts_and_from_all_job_roles_rows = [
+        (0, 10.0, MainJobRoleLabels.care_worker, 5.0, 10.0, 0.0), # All job roles have filled posts.
         (0, 10.0, MainJobRoleLabels.senior_care_worker, 5.0, 10.0, 0.0),
-        (1, None, MainJobRoleLabels.care_worker, 5.0, 10.0, None),
-        (1, None, MainJobRoleLabels.senior_care_worker, 5.0, 10.0, None),
-        (2, 10.0, MainJobRoleLabels.care_worker, None, None, None),
+        (1, 10.0, MainJobRoleLabels.care_worker, 5.0, 5.0, -5.0), # Job roles have mix of value and null.
+        (1, 10.0, MainJobRoleLabels.senior_care_worker, None, 5.0, -5.0),
+        (2, 10.0, MainJobRoleLabels.care_worker, None, None, None), # All job role posts are null.
         (2, 10.0, MainJobRoleLabels.senior_care_worker, None, None, None),
-        (3, None, MainJobRoleLabels.care_worker, None, None, None),
-        (3, None, MainJobRoleLabels.senior_care_worker, None, None, None),
-    ]
+    ]  # fmt: skip